feat(worldstereo): free aux encoders during denoise — fits memory-dmd on 32GB
The worldstereo-memory-dmd case OOM'd on a 32GB RTX 5090 because the auxiliary
UMT5 text encoder (fp32, ~23GB) and CLIP image encoder (fp32) were loaded to GPU
and kept resident through the entire ~4-step denoise, despite each running exactly
once (prompt + image encode) before the loop. Measured forward-entry footprint was
~42GB, ~36GB of which was resident weights — not activations.

Three coupled changes ("get the aux encoders off the GPU during denoise"):
- Load UMT5 + CLIP in half precision (_get_half_dtype() -> bf16 on Blackwell)
  instead of fp32. ~23GB -> ~11.5GB for UMT5; near-lossless for conditioning
  (pipeline already casts embeds to transformer dtype under bf16 autocast).
- In the DMD pipeline, bring each encoder to GPU only for its one-shot encode,
  then .to("cpu") + empty_cache() before the denoise loop. Reclaims the encoder
  weights for the whole denoise. Robust to pipeline reuse (re-homes on next call).
- _pipeline_common.encode_image now casts pixel_values to the encoder dtype
  (bf16 CLIP would otherwise raise a conv dtype mismatch on fp32 pixels).
- Dropped torch.compile on the text encoder: it runs once and is then offloaded,
  so compile was pure overhead (inductor workers + WSL2 fragility + device-move
  recompile). Block swap and VAE compile unchanged.

Validation (case000/traj2, off blockswap-only): previously OOM'd, now completes in
~2 min at ~10-30 s/it (real VRAM, not sysmem paging). PSNR vs pristine baseline =
36.77 dB (target >30) — near-lossless.

+++ hyworld2/worldgen/models/pipelines/_pipeline_common.py
@@ -134,7 +134,10 @@ class KeyframePipelineMixin:
         device: Optional[torch.device] = None,
     ):
         device = device or self._pipeline_execution_device()
-        image = self.image_processor(images=image, return_tensors="pt").to(device)
+        # Match the processor's float pixel_values to the image encoder's weight dtype.
+        # The encoder may be loaded in half precision (bf16), in which case fp32 pixel_values
+        # would raise a dtype mismatch in the first conv.
+        image = self.image_processor(images=image, return_tensors="pt").to(device, dtype=self.image_encoder.dtype)
         image_embeds = self.image_encoder(**image, output_hidden_states=True)
         return image_embeds.hidden_states[-2]
